+++ eras/mary/impl/src/Cardano/Ledger/Mary/Translation.hs
@@ -13,6 +13,7 @@
 module Cardano.Ledger.Mary.Translation where
 
 import Cardano.Ledger.Binary (DecoderError)
+import Cardano.Ledger.CertState (CommitteeState (..))
 import Cardano.Ledger.Compactible (Compactible (..))
 import Cardano.Ledger.Core
 import Cardano.Ledger.Crypto (Crypto)
@@ -79,8 +80,13 @@ instance Crypto c => TranslateEra (MaryEra c) EpochState where
 instance Crypto c => TranslateEra (MaryEra c) DState where
   translateEra _ DState {..} = pure DState {..}
 
+instance Crypto c => TranslateEra (MaryEra c) CommitteeState where
+  translateEra _ CommitteeState {..} = pure CommitteeState {..}
+
 instance Crypto c => TranslateEra (MaryEra c) VState where
-  translateEra _ VState {..} = pure VState {..}
+  translateEra ctx VState {..} = do
+    committeeState <- translateEra ctx vsCommitteeState
+    pure VState {vsCommitteeState = committeeState, ..}
 
 instance Crypto c => TranslateEra (MaryEra c) PState where
   translateEra _ PState {..} = pure PState {..}